Fix2

+++ bot.py
@@ -10,7 +10,6 @@ import datetime
 import pytz
 import logging
 import sys
-import traceback
 
 load_dotenv()
 
@@ -58,40 +57,36 @@ class Client(discord.Client):
                 continue
             scheduler.add_job(self.post_quote, trigger=IntervalTrigger(hours=24), start_date=start_date, args=[channel], id=str(guild["_id"]))
 
-    async def post_quote(self, channel: discord.TextChannel):
+    async def post_quote(self, channel: discord.TextChannel, use_db=True):
         print("POSTING QUOTE")
-        try:
-            # quotes table -> guild_id, channel_id, message_id, reaction_count
-            message_id = db.get_database(DATABASE).get_collection("quotes").find_one({"_id": channel.guild.id})
-            if not message_id:
-                return
-
-            message_channel = self.get_channel(message_id["channel_id"])
-            if not message_channel:
-                return
-            message = await message_channel.fetch_message(message_id["message_id"])
-            if not message:
-                return
-            embed = discord.Embed(title="Quote of the day", description=message.content)
-            reactions = message.reactions
-            print("Reactions:", reactions)
-            if reactions:
-                reaction_count = {reaction.emoji: reaction.count for reaction in reactions}
-                reaction_count = dict(sorted(reaction_count.items(), key=lambda item: item[1], reverse=True))
-                print(f"Reactions: {reaction_count}")
-
-                for reaction, count in reaction_count.items():
-                    embed.add_field(name=reaction, value=f"{count}", inline=True)
-            else:
-                print("No reactions found.")
-            embed.set_author(name=message.author.display_name, icon_url=message.author.avatar.url)
-            if message.attachments and message.attachments[0].filename.endswith((".png", ".jpg", ".jpeg", ".gif")):
-                embed.set_image(url=message.attachments[0].url)
-            message = await channel.send(embed=embed)
-
+        # quotes table -> guild_id, channel_id, message_id, reaction_count
+        message_id = db.get_database(DATABASE).get_collection("quotes").find_one({"_id": channel.guild.id})
+        if not message_id and use_db:
+            return
+        message_channel = self.get_channel(message_id["channel_id"])
+        if not message_channel:
+            return
+        message = await message_channel.fetch_message(message_id["message_id"])
+        if not message:
+            return
+        embed = discord.Embed(title="Quote of the day", description=message.content)
+        reactions = message.reactions
+        print("Reactions:", reactions)
+        if reactions:
+            reaction_count = {reaction.emoji: reaction.count for reaction in reactions}
+            reaction_count = dict(sorted(reaction_count.items(), key=lambda item: item[1], reverse=True))
+            print(f"Reactions: {reaction_count}")
+            for reaction, count in reaction_count.items():
+                embed.add_field(name=reaction, value=f"{count}", inline=True)
+        else:
+            print("No reactions found.")
+        embed.set_author(name=message.author.display_name, icon_url=message.author.avatar.url)
+        if message.attachments and message.attachments[0].filename.endswith((".png", ".jpg", ".jpeg", ".gif")):
+            embed.set_image(url=message.attachments[0].url)
+        message = await channel.send(embed=embed)
+        
+        if use_db:
             db.get_database(DATABASE).get_collection("quotes").delete_one({"_id": channel.guild.id})
-        except:
-            print(traceback.format_exc())
     async def on_raw_reaction_add(self, payload: discord.RawReactionActionEvent):
         if payload.user_id == self.user.id:
             return
@@ -205,7 +200,7 @@ async def force_quote(interaction: discord.Interaction):
         await interaction.followup.send("Channel not found", ephemeral=True)
         return
 
-    await bot.post_quote(channel)
+    await bot.post_quote(channel, use_db=False)
 
     await interaction.followup.send("Forced quote")
